Extract Iceberg schema accessors to root-safe module
